requirements.txt
appnope==0.1.4
APScheduler==3.11.0
asgiref==3.8.1
asttokens==3.0.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.2
comm==0.2.3
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
Django==5.2.1
django-apscheduler==0.7.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
executing==2.2.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.5.1
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
imageio==2.37.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lazy_loader==0.4
linalg==1.0.4
lpips==0.1.4
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.4
oauthlib==3.3.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
tifffile==2025.6.11
tk==0.1.0
torch==2.9.0.dev20250720
torchaudio==2.8.0.dev20250720
torchvision==0.24.0.dev20250720
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
